pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yxr-oi-cli"
description = "Simple Online Judge Cli Tool"
dynamic = ["version"]
authors = [{ name = "YeXiaoRain", email = "yexiaorain@gmail.com" }]
dependencies = [
  "requests >= 2",
  "beautifulsoup4 >= 4",
  "lxml >= 4",
  "pycryptodome >= 3.13.0",
  "click >= 8.1.3",
  "rich >= 11.0.0",
  "yxr-atcoder-core >= 0.0.3.3",
  "yxr-codeforces-core >= 0.0.2.1",
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
]
keywords = ["codeforces", "atcoder", "cli", "api", "yxr"]
license = { file = "LICENSE" }

[project.urls]
"Homepage" = "https://github.com/CroMarmot/oiTerminal"
"Bug Tracker" = "https://github.com/CroMarmot/oiTerminal/issues"

[project.scripts]
oi = "oi_cli2.cli.main:main"

[project.optional-dependencies]
tests = ['build', 'coverage >= 6', 'pytest >= 7']

[tool.hatch.build]
ignore-vcs = true
include = ["oi_cli2"]
exclude = ["tests", "docs", "build", "venv"]

[tool.hatch.version]
path = "oi_cli2/__init__.py"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --ignore=unused*"
testpaths = ["tests"]

[tool.yapf]
based_on_style = "pep8"
column_limit = 120
indent_width = 2
